refactor(#3/*): domain 별 성공 코드 global.code의 각 도메인 별 패키지에 위치하도록 변경
