--- src/main/java/org/sopt/snappinserver/api/auth/code/AuthSuccessCode.java
@@ -1,20 +1,0 @@
-package org.sopt.snappinserver.api.auth.code;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.sopt.snappinserver.global.response.code.SuccessCode;
-
-@Getter
-@RequiredArgsConstructor
-public enum AuthSuccessCode implements SuccessCode {
-
-    // 200 OK
-
-    // 201 CREATED
-
-    ;
-
-    private final int status;
-    private final String code;
-    private final String message;
-}
