--- src/main/java/org/sopt/snappinserver/global/response/code/ErrorCode.java
@@ -1,4 +1,0 @@
-package org.sopt.snappinserver.global.response.code;
-
-public interface ErrorCode extends ResponseCode {
-}
