--- src/main/java/org/sopt/snappinserver/global/response/code/SuccessCode.java
@@ -1,4 +1,0 @@
-package org.sopt.snappinserver.global.response.code;
-
-public interface SuccessCode extends ResponseCode {
-}
