+++ src/main/java/org/sopt/snappinserver/global/response/code/common/ErrorCode.java
@@ -1,0 +1,4 @@
+package org.sopt.snappinserver.global.response.code.common;
+
+public interface ErrorCode extends ResponseCode {
+}
