+++ src/main/java/org/sopt/snappinserver/global/response/code/common/SuccessCode.java
@@ -1,0 +1,4 @@
+package org.sopt.snappinserver.global.response.code.common;
+
+public interface SuccessCode extends ResponseCode {
+}
