+++ src/main/java/org/sopt/snappinserver/global/response/code/photo/PhotoSuccessCode.java
@@ -1,12 +1,12 @@
-package org.sopt.snappinserver.api.category.code;
+package org.sopt.snappinserver.global.response.code.photo;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.sopt.snappinserver.global.response.code.SuccessCode;
+import org.sopt.snappinserver.global.response.code.common.SuccessCode;
 
 @Getter
 @RequiredArgsConstructor
-public enum CategorySuccessCode implements SuccessCode {
+public enum PhotoSuccessCode implements SuccessCode {
 
     // 200 OK
