+++ src/main/java/org/sopt/snappinserver/global/response/code/place/PlaceSuccessCode.java
@@ -1,12 +1,12 @@
-package org.sopt.snappinserver.api.curation.code;
+package org.sopt.snappinserver.global.response.code.place;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.sopt.snappinserver.global.response.code.SuccessCode;
+import org.sopt.snappinserver.global.response.code.common.SuccessCode;
 
 @Getter
 @RequiredArgsConstructor
-public enum CurationSuccessCode implements SuccessCode {
+public enum PlaceSuccessCode implements SuccessCode {
 
     // 200 OK
 
@@ -17,4 +17,5 @@ public enum CurationSuccessCode implements SuccessCode {
     private final int status;
     private final String code;
     private final String message;
+
 }
