+++ src/main/java/org/sopt/snappinserver/global/response/code/user/UserSuccessCode.java
@@ -1,12 +1,12 @@
-package org.sopt.snappinserver.domain.photo.controller.code;
+package org.sopt.snappinserver.global.response.code.user;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.sopt.snappinserver.global.response.code.SuccessCode;
+import org.sopt.snappinserver.global.response.code.common.SuccessCode;
 
 @Getter
 @RequiredArgsConstructor
-public enum PhotoSuccessCode implements SuccessCode {
+public enum UserSuccessCode implements SuccessCode {
 
     // 200 OK
